fancier .MigrateFrom / .MigrateContent
for widgets with a fixed number of dependencies of different types (eg MainWindow with it's "menu" and "content" dependencies), where you might want to intelligently swap in and out the dependencies (vs. a full remove and refill of many identically-typed items, as in box layouts) ... there is now less boilerplate to write, and it's more clear what's going on.