taming the complexity
OK, so maybe finally an example of what 2-way data binding will look
like WITHOUT exposing the full internals of a given control

because DatePicker still has its own 'raw' field that's not exposed
and yet it's able to expose 3 different states which can also be
intelligently set from outside.

the question is, is this a true improvement upon the old MVC for complex
validation?