solves the crashing problem, but ...
strangely when we changed the disposeTree code + IBuilderNode interface
(to add an 'isOwned' flag) to avoid disposing things owned by Qt windows
as an ancestor, everything was leaking the callback functions assigned
to signals.

now have have simplified all that simply to avoid disposing Dialogs that
are (presumably) owned by a window ... but nothing's being leaked? huh?

a crappy solution for now but need more time to think about this whole
horrible node hierarchy and how to potentially simplify it, perhaps by
defining it (with default interface methods, or something) in C#
instead.